(console/core) auth-chld dms ignore-race
Thanks to Amadan-Pobretano-LaBoon for reporting.

This was only really visible from the context of authoritative launch
of a client that thereafter quite voluntarily. This is the case for
decode when you watch a clip to the end without enabling looping, and
when a terminal is exit through 'exit'. Normally, durden, pipeworld et
al. just flat out kills the client so this scenario would not occur.

When the primary segment shuts down and sets the dms, it was not chained
to arcan_frameserver_free correctly, as well as waitpid not handling the
case where the monitored child process do not exist. The later check is
one that still should be considered for removal, or making the contract
for authoritative/trusted launches that if the primary segment pid dies
we sever the connection.

This also fixes a problem with console.lua not respecting origo_ll
texture coordinates (this old story), e.g. when launching an arcan_lwa
client from the console.