(tui/shmif) add cell_w/h to init block
With some seemingly minor changes to displayhint propagation,
the terminal started having issues with using the wrong grid size.

This was an artifact of a display-hint with the proper cell
dimensions almost always appearing after the first frame delivery.
Since it is now likely to know the cell dimensions after preroll
synch, add them to the shmif_initial structure and call as well.