(decode) add _img support
This is a small step with large consequences. The intention is to move
all image decoding into the decode frameserver and have the parsing
there be least privilege when it matters, yet allow the main engine to
retain its old way when it doesn't.

To achieve this we add a new proto=image argument to afsrv decode, and
using the bchunk mechanism to announce supported types and transfers.

This lets us then write a hookscript that overrides load_image_asynch
with calls into afsrv_decode, then using a combination of
resample_image (current frame to new store),
restore_target (send the file) to implement the actual decode + load.

Another nice detail with this setup (testing with nanosvg for the time
being) is that it is also useable for rerasterization on density change
for mixed DPI outputs.

FossilOrigin-Name: 01d228f2b91275b243b335f2de4c692f1c3a4ffbefdb183646828b5ed0d2065a