(conductor) workaround for kernels with low tickrate
This fixes #197. The problem was that the platform 'yield to conductor
during vsynch block' would stall in large blocks on certain OSes, notably
OpenBSD. This comes from the default kernel timer resolution being 100Hz,
making most tricks short of spinning practically useless. The solution
is to query the timing behavior from the platform to determine the minimum
reliable timestep if the kernel is not tickless.