(egl-dri/lua) expose vrr/direct_fbo dynamically
Direct_fbo is likely not to remain a default-off runtime-cfgtime-on
but until the last rather tricky swap-order problems have been solved
(which they should with ldr/hdr scanout) it should be provided as a
manual control that latches onto modesetting.

Enabling vrr likely won't have any sizable impact as the other conductor
controls are not there in terms of explicit fencing and slew rate
control, but should be available for experimentation and measurements.