(core/lua) add oob nbio- callback
This adds a :data_handler to the open_nonblock i/o userdata. This will
trigger when data can be read or written (depending on mode) once. The
callback will often come when other I/O and synchronization activity is
blocked, providing a slow trigger for doing basic manual i/o while
waiting for gpu-blocked operations to complete.

The plan is to expand this with deferred buffer transfers - triggering
the callback when the a transfer has been completed, as well as re-use
the interface for an asynch glob_resource.

The other internal use is to have a way for the conductor and the
platform layer to have I/O triggered descriptor state changes instead of
vsynch or timesleeps.