(shmif/server) split out execve bits of handover
First in a long line of shmif implementation refactors in order to one
day make it more portable to non-posix environments (like it used to
back in the day).

The handover- exec implementation is split into two parts, one that
extracts the necessary primitives from the context structure and event
metadata, and the other that actually builds the environment.

This is also added to the shmif server implementation so that the
argument form where a specific binary+env is to be executed and bound
into a shmifsrv_client structure. The use for this is to have arcan-net
allow a -exec mode where a newly authenticated a12 client gets spawned
and assigned an arcan-shmif compatible local binary.

FossilOrigin-Name: e455d63c44e6a5caf3e81ae56b5e09a31f3bfe79c409763e08f88d41d3c0b251