(a12) draft -l -exec mode
This extends the previous commit with the command-line argument
processing and setup for the 'reverse' (or forward depending on how
you see it) mode. The point is to be able to do say:

    arcan-net -l 666 -exec /usr/bin/afsrv_terminal

and a remote a12 client should be able to connect and get the client
back. This works in reverse from the normal setup:

    arcan-net -s test remote.host 666
    ARCAN_CONNPATH=test afsrv_terminal

    (remote)
        arcan-net -l 666

or the short form:

    ARCAN_CONNPATH=a12://remote.host:666 afsrv_terminal

where afsrv_terminal connects to 'test' (local), which creates an
outbound connection to remote, that then forwards it to whatever is
the ARCAN_CONNPATH display server that will render there.

This was built on how clients worked in X, hence the display 'server'.
What we want in other contexts is the opposite, the application is
served to a remote viewer.

This patch adds most of that, e.g. the connection and negotiation
goes through, but then stalls for some reason. arcan-net can't be used
as a viewer, but this is what afsrv_remoting is for:

    ARCAN_ARG=protocol=a12:host=remote.host:port=6680 afsrv_remoting

But for some reason the events does not get matched back and forth
correctly yet.