(evdev) FreeBSD + evdev fixes
Previously the build switched to wscons even if evdev was explicitly
specified. This change should permit evdev to be used. It is still
not the default, as some minor work is still needed:

This is still rather untested and missing hotplug. For the later we
still need to do something like read from the devd socket to know
when we should issue a rescan.

FossilOrigin-Name: 4fcb9f7d25a5c1aad214a3bef5bdfb90b11b7759440eb3965d0722fe53c2c04a