(shmif) handover-exec CLOEXEC for shmif-server
This was one major reason the -exec mode to arcan-net didn't work. The
way a new server is setup and handover-executed meant that the control
socket was being created in close-on-execute, causing the new client to
fail to open the socket.

There is still some other reason as to why Xarcan (specifically) doesn't
work in this execution mode.