(tui) assume glyph presence on unknown fonts
It is possible to run glyph presence queries on a context that has
yet to actually setup font state, either from not receiving them or
before the first possible frame. This would cause a null-deref with
alternate shmif- implementations (e.g. libshmif-server). For that
edge case, assume that the glyph is actually present.