(egl-dri) mask out hint_direct, reorder logging
This reflects changes to allow the direct-scanout-fbo option to be
set dynamically, both for testing and vrr-hdr-... toggles.

Since there are so many conditions that can make the direct scanout
infeasable, the logging had to be extended and to make that not spam
on debug builds, the argument evaluation order changed somewhat.

FossilOrigin-Name: 42b0475989b62b1ad9e9a1b416224aa243985f36adfdb8fb603035bf7dabeb60