(core/egl-dri) dirty pipeline and decay tracking
There has been a regression in how the chain of rendertargets
communicate if there are pending transforms (== can always produce a new
frame), if there are invalidated backing stores (should produce a new
frame), if there is nothing new and if there is nothing new but the
platform needs new frames to be produced regardless.

This is a complex part of the pipeline as is, and still incomplete as it
is binary and not per rendertarget. It will also start to matter more as
the horrors of HDR composition unfolds.

FossilOrigin-Name: fd14f5ca93b32db67b14a1eee74bf727c8c97a3177895e23ab080e27b17b8bbc