(sdl2) attempt at resize-hdpi fix XXIV
There are several issues with how SDL2 windows resize when inside
another windowing system, and depends on the level of HDPI support as
this is not something that cleanly translates between say, OSX, X11 and
Wayland because of insistance on 'Scaling' hacks.

Even if we announce HDPI support, there might be an internal scale
factor applied to the window sizes and so on by SDL itself - and there
is no function to expose the state for that. We can derive it by looking
at the ratio between logical size and window size and then use that to
figure out which dimensions we should actually switch to while retaining
density aware rasterization.

With the comes an ugly edge case in floating point precision, window
size and a possible WM feedback loop in that we pick a size that fits
then get a new hint that is rounded to fit some WM criteria (e.g. %2 ==
0) which causes a new resize and it keeps oscillating.

There are no event cookies available to detect this, so instead a
time-cap is the closest thing to a generalized solution that can be
applied. Actually testing this basically takes drag-resize interactively
on the harder platforms (Quartz, Wayland). Especially with Wayland the
behaviour across compositors is incredibly inconsistent -- perfect storm
of poorly specified behaviour, overly specified interfaces and zero test
cases.