(egl-dri) mode-buffer format switching for comp.
This should make the 565 to 888 to 10bit switching practically instant
but with a caveat that this might work poorly on other/ older EGL
implementations, but this platform is so tied to MESA anyway.

The caveat is documented further in the egl-dri/video.c, but the gist is
that EGL legacy required a configuration to limit supported surface
formats during context creation. Rebuilding contexts meant losing state
which was anything but pleasant (treated as GPU failure internally).

With EGL_KHR_NO_CONFIG_CONTEXT we can forego most of that, and possible
dump all the device/display context switching back and forth as it makes
little sense anymore.

Most of this are moot with the regular FBO 'direct' scanout mode, but
that one always require an EGL fallback for edge cases where we want to
compose with a shader or texture coordinate transform.

FossilOrigin-Name: 530ef8ec293756e6bf0e4b6fc178790cd3a2a5c15ca7293a18adf7d8a005e9a9