(egl-dri/agp) float fbo allocation changes
The format and context setup was incorrect as to how the config was
specified in order for a proper EGLconfig with a matching low level
visual to work. This should be more 'right' but still won't produce
the right EGL context.

Direct mapping (HINT_DIRECT, ...) has generated both crashes in the
kernel, the GPU and arcan itself sporadically so we are in fairly
unexplored territory here.

FossilOrigin-Name: 5b608d79e1443f154241ba68e458270e0d4d3269683657a2cff0f68219b9310f