(lwa) rework synch logic
This reworks the lwa platform to be closer to variable refresh by
default, but mainly as a hack to not be artificially locked to 60Hz.

This is a temporary measure, it is just that better primitives take much
more plumbing and the lower level infrastructure is still in flux. The
problem is not with engine render/processing logic itself - but rather
the lack of reliable sharable synch primitives and presentation time
logic:

 1. the server needs to communicate (and update) deadlines
 2. the client platform needs a multiplexable descriptor for when
    a display can be synched onwards.
 3. the processing needs to be able to lock on a rendertarget basis
    so that other rendering / processing can take place.
 4. the conductor needs to be able to apply its event processing
    pollset with that of locked displays.
 5. rendertargets bound to subsegments need a FBO based sharing/reblit
    mechanism.

(3) has larger consequences due to the _preframe/_postframe synch entry
points as it makes incrementally less sense.

A big caveat to this fix is that multiple displays (mostly an edge used
for testing display hotplug logic without being in the native engine
context).

FossilOrigin-Name: c6f1b8f4a649295d4828f5744cb5c181657eb2264da18e5b71c302376081cf73