(a12) evict miniz
With the zstd migration finished with (9f1d19e) there is no reason
to keep the inferior zlib and the horrible vendored miniz around.

Kill it with fire.

This might break som very custom a12 build, but that already happend
against the packaged versions way back when so better to get it done
now right before a new tagged version.