(builtin/mouse) add a simplified mouse_setup()
The argument processing for mouse_setup was always.. du^H exotic.  While
this is due for a mouse2 that actually returns a ctx in order to have
multiple ongoing state machines and so on, band-aids are still needed.

This simply checks the arguments for the common patterns, and if they
are not provided, fills in with some sane:ish defaults. mouse_input also
takes on another branch on argument type to route through
mouse_iotbl_input.