(lua) prepare for user-defined namespaces
This will not make it into the next release, but the API changes will
need some time to propagate. This adds a string- typed 2nd argument to
glob_resource, allowing arcan_db to act as a resolver for namespace
globbing (virtualizing storage).

The next step to that is to have a prefix in load_image, open_nonblock
etc. which allows a namespace_tag:/path/to/something