(a12) most of the scaffolding for backpressure
This is one of the big ticket items for a12, and that is constraining
frame production based on congestion feedback. Most of shmif is designed
to explicitly handle and test that, but the a12 layer has so far been
very lenient and consume frames on the producer side regardless of the
network state to the consumer.

FossilOrigin-Name: cdcaa907263c469b5db86e85b6ea9b597572d74e52aca6817dca0802121f4926