(a12) measure vframe backpressure
This adds the painful tracking of encoded and submitted video frames
versus acknowledged ones. After adding an accessor to the data, it
should be enough for some basic congestion control and adaptive
bitrate selection.

Those parts are kept out of the a12 implementation itself, and will
be added to the -net tool (as well as afsrv_***) via the vcodec
selection callback. It makes more sense to have the heuristics there
as other actions might be possible (defer congestion to the source,
buffer locally and have a timer, ...) though there are some grey
area with encoder configuration in particular as that is still the
responsibility of the a12 implementation.