(lua) cursorhint event message format
The filter used to sanitize-push the message for cursorhints excluded :
and digits, making the 'last resort' warp:, hot:, etc. state messages
lose information.

While the proper way still is to have a cursor subsegment and use
viewport anchoring for warping and hotspots, this is still needed for
some edge cases, like VMs where the cursor is force-composed.