(a12) keystore/ephemeral mode breaks from -1
The previous 'forgotten branch' squish managed to mess up quite a few
things with both key retrieval and setup, especially for the round-trip
ephemeral mode. This should be a much better point for further
validation. It also adds an explit getpass() like query for the pkauth
mode, as well as explicit controls for disabling the ephemeral-rt mode.

Any controls that reduce the security can now only be toggled via long
form arguments so that it is more visible that it is going on.