(a12) auth event drop-race
This is a rare but fun one. By default, a12_unpack will always consume
all the supplied bytes in the buffer. This has the edge case where the
incoming buffer contains the last necessary bytes for authentication AND
relevant main-loop events.

If the API consumer needs to do something between authentication and
loop (e.g. map a shmif context on a channel), chances are that any
trailing bytes will be decoded but results discarded. Practically this
would really only happen on events due to intermediate buffer sizes, but
that could carry something like a REGISTER event, causing the source to
stall indefinitely.

The workaround for this is to add a latched state to authentication so
that if there are remaining bytes to process when the state machine
transitions, move that to an intermediate buffer that takes precedence
on the next a12_unpack call, and apply a flush-unpack (0 incoming bytes)
form.