(a12) fd leak / credentials not copied
This masked the issue with -exec and -listen modes when combied with
-t (which is not the default, but helps with troubleshooting). The
problem was that the opt->secret fields is cleared on consume after
it has been used. Since the default is forking, that was fine since
the allocation would naturally differentiate.

In non-forking / single connection mode on the other hand, that lead
to the structure being reused but with different members for some
key values, preventing future connections.