(a12) respect the origo_ll flag
This should cause lwa- and other clients to render 'right side up'.  To
save postprocessing at some key positions, it is normal that a shmif
source will mark that the image is stored with the y-axis inverted.

This is a legacy from how GL managed its textures, and especially from
where FBOs (framebuffer objects) were drawing into a colour attachment
(texture) where the coordinate system has 0,0 in the lower left corner.

This is in conflict with scanout engines etc. where 0,0 is normally in
the upper left corner.