(arcterm) expose ARCAN_STATEPATH
This has previously been blocked from spawned terminals, but since it is
currently useful to define / reach the arcan-net keystore, it should be
exposed onwards for the time being.

Eventually the namespace env. pollution should really just be moved to a
dynamic lookup function in shmif (which at first can return the env) as
that allows for a migration path where we can query this dynamically.