(a12) add intent to HELLO
With the -exec mode comes the conflict where the primary connection for
the inbound OR outbound connection might be a shmif source. In the case
where we then connection an outbound connection that is a source to a
listening endpoint that is also a source, there is a conflict of what
the connection is supposed to be used for.

This patch attempts to work around that be expressing intent in the
HELLO messages, and subsequently marking -s and -exec side as source and
regular ones as sink.

When authenticating a HELLO, each side then checks so that the remote
is_source differs from the local one, and if not, terminates the
connection.

Since this would break existing clients, a backwards compatible
'dontcare' value remains the default, and the connection continues as
normal in that case.