(a12) add probe- state to HELLO packet
This prepares for two things, one is a mode where an authenticated
client will just terminate after authentication as a means of probing
what the other end can supply. This is important to distinguish between
SOURCE, SINK but also the coming DIRECTORY mode.

While the ink hasn't dried on the ideas for DIRECTORY, the plan is to
use it for discovery, NAT puncher / proxy, as a state-store/backup and
as an appl delivery/transport.

For the appl case, the statepath expands with a possible state store/
restore for the authenticated user and the APPLBASE is used to basically
tar + checksum + sign -> binary-stream -> unpack into cache ->
chain-exec through arcan_lwa.