(decode) add protocol=list
This just returns a message with the set of supported protocols in the
current build. This is mainly useful to WM- like file pickers and so on
for determining a strategy for handling thumbnails and previews.

FossilOrigin-Name: 07caaf3046c484b9dac1d613130bdcaef9de467a54edd3c96db10fff2716ea81