(tui) readline- add basic history support
We should perhaps implement the C-f using current workstr as search
pattern for building a history subset. With the more interesting or
powerful way of achieving that takes regex engines the argument, it
might be better to just leave this to the user as an input binding that
explicitly samples the current readline and swaps out history.