(lua/build) first steps for splitting _lua.c
This has been a long time coming, and will be a slow and painful
process. Basically the _lua.c behemoth is to be split up into smaller
logical translation units. This patch takes care of the VM management
bits that deal with crash recovery and calling into scripts.

Some rough draft of what the discrete steps should be:

 - moving other Lua support functions (type conversion)
 - event to entry-point mapping
 - each 'category' group
 - document- generation scripts with some checksum/tracking that
   any change to the layer has corresponding updates/errata in
   document.
 - abstracting the lua-VM integration functions into overridable
   symbols

The last stage is perhaps a bit surprising, but it is basically the
precursor for an ALT- bytecode format. The intent there is to be able to
use the high-level API from other languages in external processes.

FossilOrigin-Name: 8f5093b1c4dae3a28598a85d39cd8f470aa46a5b85e1ea91d31aaf45edb5a8f3