(encode) switch to non-deprecated enc
This is a quick cut-n-paste job to ensure that things build on ffmpeg5
and above that has removed symbols and members we relied on. It is
likely that this will break encoding entirely or when audio/video is to
be muxed.

This also splits out the ffmpeg bits in order to make that dependency
optional for fsrv-encode later, and possibly add a less license-dep
heavy alternative. Ideally the entire encode_ffmpeg should be rewritten
as there is so much ugly in there following the evolution of ffmpeg APIs
over more than a decade.

FossilOrigin-Name: 4f9410eea2e14396fa062ec6f1406c8b57b97ab1d41ca6c6a044b6fa61958356