(tui) handover- mapping to virtual window
This creates an unbound tui window (usable only for rendering) to be
used on on_subwindow when receiving a HANDOVER segment. The use of this
is to be able to control embedding position post- handover.

An open question still is to influence embedded handover presentation
parameters, with a possibility being to reuse the segment_token and an
PROXY->TARGET event.

The flow would be something like:

    cl: request handover
    srv: allocate-push handover
    cl: take-handover, remember segment_token, handover_exec
    cl: queue-external PROXY->segment_token, queue TARGET event
    srv: mark segment_token, set proxy-dst, forward TARGET

Normally any TARGET event would've been discarded in that direction as
per the normal filter. It is a powerful mechanism, but some problems
with queue contention (failure needs to be notifiable) and should
probably be restricted to handover- segments. Events that are
descriptor-carrying also need to be blocked still.

FossilOrigin-Name: 3bed35ca67a63a66aefd999b75754695f1c6c28a2e9f2ecf94bd8f622095407a