(egl-dri) synch display - rtgt framecookie
This line must have been accidentally dropped in the past. The
consequences are that a mapped rendertarget always makes the
display being considered dirty. This would not matter unless the
same rendertarget is mapped to multiple diplays, hence why it was
not spotted.

Another subtle side-effect, however, is that the display and
rendertarget are always marked as out of synch and ready to update.
The problem then is that if the display is marked as suspended or
off, this would prompt an update to be pushed and the conductor
set to wait for the next vsynch - which will not be triggered
since the display is off, causing a full core spin.

FossilOrigin-Name: 22902d53e2908456c4027f334acbdd6ab4352dacfead57c7d435c636613f1f84