(tui/arcterm) synch lua-tui bindings+lash
This merges the latest work on the lua-tui bindings and how they
integrate into afsrv_terminal with ARG=cli=lua. The first crude outer
usershell chainloader is also pulled in.

It seems more likely that the 'original' afsrv_terminal + tsm will
eventually be deprecated in favour of just having lash with a vt compat
mode implementation - the cost vs safety tradeoff is favorable, and a
lua implementation will be much more maintainable vs. the current set of
hacks.

Ultimately this should mean that the entire tsm + tsm_screen can be
dropped entirely (-10klocish) and all tui screen operations should be
faster.

The current blocker outside of the missing vt- state machine is
basically that there is no low-level 'line-mode' widget state yet, and
that the pty setup code needs to work differently on OSX due to their
multiplexing not working on ptyx devices.