(tui+lua) incorrect subwindow error handling
The error code was being set wrong making it impossible to distinguish
between the fd pollset having a broken descriptor, or the context set
being broken.

Reworked that and added compaction/cleanup code for the meta. This will
still not invoke the tuiclose GC, it is possible that the parent ref
needs to luaL_unref for that to happen.

There is also a subwindow mapping race that might be on the arcan end of
things, but sometimes the open #1 new (lash/cat9) command will have the
new window get the wrong (fallback) font rather than inheriting from its
parent.