(shmif+tui) handover_pipe call
The handover_exec function only ever ensured that stdio went somewhere
and controls for detaching to /dev/null. Other cases you want to provide
additional descriptors to be inherited, or create pipes that map to
stdio and is returned. The later case turned out to be practically
necessary for the tui-lua bindings.

An extended function for this purpose has been added and the internals
refactored to build on top of it.