(tui) readline: add insertion prefix control
this lets the caller define (use with _suggest) a prefix that will
be added when the user picks something from the list of suggestions.

the need for this is when you want to present a set of items with
the prefix stripped, but have the end-result string contain it.

example:

    cd ./c<TAB> -> [close, close] <picks close>
    set prefix to cd ./ and the substitution insertion mode will
    make it cd ./close

not having this would make subst. only give 'close', append give
cd ./cclose and word give cdclose.

FossilOrigin-Name: e00fd16b523ca238a2cb95e7325d2b7a2a382bba3eaef5f8cdfd14ad466a67dd