(tui) rl: verify on enter, preserve in_completion
This makes it so that tab completion requires an explicit cancellation
via ESC not to be continously invoked as the user navigates. For this
not to get stuck on the last entry but rather double-enter to commit,
verify gets called immediately again on enter/right-step, and it becomes
a commit/finish step if the new completion set is empty.

FossilOrigin-Name: ed17a544f7514ce586ed1160ed13e39d6af2e617c1128bc14a94d6b45582ea89