(tui) add api control to suggestion-popup
Prior to this patch suggestion completion would only be triggered
when the user explicitly pressed TAB. Sometimes the application may
have better contextual information as to what the trigger would be.

This patch adds a function to explicity enable or disable the popup
without user input. For the Lua bindings this does not add any new
entrypoints, just an overloaded :suggest(bool).