(tui) readline world completion space on !cursor
Word completion had two problems, one is that prefix was always
ignored, and the other that it is implemented as delete_word +
insert. This will strip the whitespace from that word as well,
which should be reinserted unless it is at the start of the line.

FossilOrigin-Name: 6cdddfb8fdc2f1494076a11891840e5dc13300312ea6e559e8bdbcbe39087db1