(arcterm) option to propagate stderr (keep_stderr)
For some cases it is useful to let the stderr of the terminal propagate
into the new child and not use it for the emulator itself. This option
will do that as well as re-inject if the emulator is reset. It also
fixes a case with handover_exec keeping descriptors to the inherited
pipes, blocking SIGPIPE or corresponding -1 EPIPE from propagating.

FossilOrigin-Name: 9ec760185d1ff106f10f4e00f863e65865f5ff411325b536ef6a639745a2b98e