(lwa) workaround for initial livelock
There is a rather serious timing issue around the initial resize storm
when the client also switches to handle+vsignal based stepping where
the first STEPFRAME is lost or never emitted at all. Since the issue is
rather limited and the entire setup around this need to change as soon
as we can portably use fences for this, just add a short timeout to make
sure it recovers. The worst that can happen is tearing until steady
state happens a frame or two later.