(distr) console/keyboard - add repeat-rate
This adds automatic repeat-rate controls to the keyboard support script
and enables its use in console.lua. The reason for this is mainly the
inconsistency between how platforms deliver this information and that it
seems less and less likely to be cleanly solved at that level.

FossilOrigin-Name: 819eaa6fd61c43c955dd8742f54613555cdcbc2cf5e75812cf7b6debd94176b3