(net/lua) re-add net_open
net_open interface has now been decided upon as something similar
to target_alloc, but with an intermediate segment_request step for the
actual primary connection - with the primary connection acting as both
metadata container for statistics / troubleshooting as well as a
mechanism for reconnecting.

    net_open("@laptop",
    function(source, status)
        if status.kind == "segment_request" then
            accept_target(w, h, real_handler)
        end
    end)

this is intended to be used with net_discover or net_browse (yet to
be added). net_discover for finding previously known, net_browse to
connect to a directory.