(build) evict vendored ttf
This was a long time coming, though there are a step or two left - the
culprit is mainly shmif-tui which strictly doesn't need to know much
about the fonts except which glyphs are available or when ligatures are
supported and the GSUB tables need to be processed. A possible 'way out'
would be to eventually invent a compacted meta format that only contain
this information.

relatedly - the modified arcan_ttf.c (comes from SDL_ttf) is starting to
show its age and most of the features we added have been included in
upstream SDL (although it has grown to twice its current size).