(readline) opt for no-bind, position ctl and fmt
This adds a control option for disabling all builtin keybindings
for those that want to deal with that themselves. It also adds a
function to set the cursor position, as well as a list of formats
that apply after each codepoint offset.

FossilOrigin-Name: bbc75162c2ae1622af286efc9456a51a8df8e52cafb65bcf293265954bac0cc6