(core/tui) extend tpack to cover cursor colour
This was an oversight from quite a while back - the cursor colour was
actually hardcoded. This change adds an extended cursor header to tpack
(will be used for more things) that can convey the current cursor color.

Hooks have been added to have separate cursor renderpasses, the main
intent for this will be to let shader-/composition- stage determine the
actual cursor shape, preferably with animation.

The TUI APIs themselves doesn't make it convenient, but tpack is
prepared for having multiple cursors and cursors of variable length.

The vstore engine side is also extended to start covering a local copy
of the tpack buffer itself, which is going to be useful when we switch
to shared atlas drawing.

The rough plan for this:

1. complete custom cursor using a proxy vobj with shader and callback
   that gets drawn when the surface itself is drawn (even if the
   storage is shared elsewhere)
2. with map_video_display - we need something else for the cursor
3. update image_access_storage for sampling the tpack state
4. have tpack rasteriser populate binpack for font-context
5. change rasteriser to sample from tpack and build command buffer