(a12) tweaks and fixes to --probe-only (1/2)
this fixes some of the server end issues with probe mode, along with
a net -l host@ -exec argument parsing error causing null-deref. The
other side of probe fixes will come separately. The problem was that
the server end prematurely rejected the probe-role or treated it as
a normal connection causing unnecessary -exec calls.