(net) plugging in directory mode
This adds most of the scaffolding and protocol changes for the app-
end of directory mode. This means arcan-net will be able to serve
an applbase dir, and, later on, act as state store.

The last stage to this feature will be to allow register /
notification for pub/sub discovery of other clients and transparently
relay the net_ functions used by the chainloaded arcan instance.

Some things being considered still are also a proxy/relay mode for
NAT hole punching and for actively tunnel channels between clients.

It is still quite unsafe and arcan-net really needs an actual config
format and more reliable state / cache store now, but it is a start.

For quick mitigation, a hookscript that disables system_load of .so
files is a start - then forcing certain namespaces, blocking fonts
and so on.