Implement GNU install directory convention (#250)
* (build) Implement GNU directory conventions

With this patch, many hardcoded paths are now substituted by GNU conventions, as
documented in the references below:

https://cmake.org/cmake/help/v3.23/module/GNUInstallDirs.html
https://www.gnu.org/prep/standards/html_node/Directory-Variables.html

Caution: this was written and tested only for Linux and in Linux; however it
should be extended to other Unix-like systems.

* (build) Implement GNU directory conventions on BSD systems too

This commit removes some hardcoded paths on BSD systems. Now hopefully all
Unix-like systems will respect the GNU directory conventions.