(lua/platform) namespace refactor broke some formats
The wrong set of flags / options were being passed when opening
some files after the reworked user-namespace rules. This caused
fonts being passed to clients to be ignored, causing tui clients
to default back to the built-in safe one. Worse still, the open
flags would also be ignored, leaking the descriptors to children
due to a lack of CLOEXEC.