(net) server --directory cache appl-tars
This moves the app-scan stage to be part of setting up the server,
and each new 'forked out' client will share the same cache. This is
still not the desired ideal, a tempfile (or permanent) sqlite cache
database would be the next valid step, and then have some monitor
thread / control signal for repopulating.