(core/net) refactor monitor mode, state synch
Another big item covered, the monitor mode has been reworked in a
way that allows an external chainloader to control more behaviour
of an arcan process. This is leveraged by arcan-net when running
an appl to collect crash samples and save/restore state.

Missing still is that the server end does not accept the state or
log, initial version should just store it as b64(H(kpub, applname)).

Later with a proper database and some configuration for non-appl
bound state stores there is a need to revisit this, but PoC for
now until the easier netappl dev bits are getting sorted.