(net) wire-in test directory state store
This extends the keystore with the ability to store associated data
indexed based on the public-key, along with the plumbing from having
a client bstream request have its blob+debug map into matching tags.

A client requesting an appl from the directory will also first have
the server check the state store for the [key, applname] pair and
send the state first (if it is there).

While mostly untested and some minor things to hammer out, this
should permit config persistence across machines for the appl case.