(net) bcaching rampup and fixes, basic compression
With this patch the local cache + cancellation is actually working,
and the sending end has added a rampup so that it won't queue /
burst until we've seen a later serial to give the cache enough time
to have an effect.

It also adds the per bframe-chunk compression and zstd setup.