(tui/lua) add writeattr_to
This is a minor performance nit for cases where one want to update
the attr, such as recoloring or highlighting - without reading back
or otherwise track the actual codepoint contents.

It adds a function that supports updating only the attribute portion,
and is added to the lua bindings as an overloaded form of write_to.