(encode) ffmpeg 5.0 regression fixed
The previous push to this was mostly a placeholder to get rid of
deprecated warnings and did not exactly work. This time around most
issues should be sorted, and while the code remains ugly as sin,
it is at least back to a tolerable / usable state.

FossilOrigin-Name: fe02407a4af238ee97a106fb854f6899f60661975929f3028c370374adf10477