(tui/alt) nbio: allow cancelout from read(fn)
The fn-callback and table versions of the read function both suffer
if the source is inifinite and faster than the processing, causing
a livelock. Normally the data_handler is used for that, but for
'foreach line' kind of contexts the other form mixes better in some
contexts.

To avoid breaking current use, this adds the behaviour that a true
return from the callback:

    a:read(function(line, eof) ... return true)

will cancel processing until the next call. Similarly, the table
form now has a hidden key member:

    local tbl = {read_cap = 10}
    a:read(tbl)

will only read read_cap elements.