(tui/tui-lua) wndhint embedding scale fixes
The wndhint parts were left in an incomplete state last release.
This should add the missing 'core' parts, though there still needs
to be some testing done in durden and cat9.

The problem is:

1. tui client requests an embedded handover surface at x,y
2. tui client receives and hands over
3. external client begins drawing
4. tui clients moves and resizes the handover window to be
   'iconified' (scaled)
5. WM respects this and picks the best fit.
6. tui client now wants it to be drawn best client determined fit.
   within some constraints.
7. WM forwards the hint.
8. client resizes and draws.
8. WM gets the new frame, needs to communicate this..

with a bonus of 9. WM decides to detach and external client should
not try and compose (basically ignore) it until it dies or reattaches.

Both need the same thing, basically a way to route displayhints to
the proxy window. To do this we add some basic parent/child tracking
and in the displayhint with a matching subsegment token, redirect
the dimensions to the handler of the proxy window.