(tui/readline/lua/alt) + readline completion hint
This changes the auto-completion presentation to also include an
optional hint. To avoid breaking API or introduce new entry points,
it is added as a bit setting to the hints provided when adding a
suggestion.

If that bit is set, the suggestion table is treated as a doubly
terminated (value\0hint\0) string. In the Lua bindings, this is
handled with an extra field to the suggestion table:

    suggest({"a", "b", hints = {"(a)", "(b)"}}

which would show ">a (a)\n b (b)". Applying the completion will
only have "a" being inserted.

also fix first history stepping being one "too far" and synch
nbio changes to engine/alt.

FossilOrigin-Name: 64bba7305c2c987d8478a5a39b7bb2da8ff9ad25274e4e6f5f466f0ff9df48bf