(tui) add border rasterization
To avoid the use of 'line drawing' characters some mechanisms are
in place in the TPACK format. On is to have a 'skip' attribute
indicate that it is meta contents, while another is to use border
attributes. These have not yet been respected by TUI itself and
now is the time to do so.

This patch adds basic TPACK properties for border drawing.

Some open issues are if the border width should be configurable,
or how it should otherwise use font information to size or position
the border when there is 'space' vs 'nil' vs 'character' in the
cell where the border attribute is set. The foreground color is also
used as we have no other position to actually set that. Since we are
heavy on the 'bytes per cell' count already, adding more would be
unpleasant unless going for stealing some of those precious
bitpatterns used for terminal instructions in lower ASCII.

One option would be to center the drawing when it is empty,
otherwise set thickness based on cell height and round the corners
when TL, TR, BL or BR are set.